+++ pkg/backup/create_remote.go
@@ -18,7 +18,7 @@ func (b *Backuper) CreateToRemote(backupName string, deleteSource bool, diffFrom
 	if err := b.CreateBackup(backupName, diffFromRemote, tablePattern, partitions, schemaOnly, backupRBAC, rbacOnly, backupConfigs, configsOnly, skipCheckPartsColumns, version, commandId); err != nil {
 		return err
 	}
-	if err := b.Upload(backupName, deleteSource, diffFrom, diffFromRemote, tablePattern, partitions, schemaOnly, resume, commandId); err != nil {
+	if err := b.Upload(backupName, deleteSource, diffFrom, diffFromRemote, tablePattern, partitions, schemaOnly, resume, version, commandId); err != nil {
 		return err
 	}